Refactor the code, split classes and fixtures
We now have a dedicated modules that provide that high-level implementations
and abstractions (e.g. AkonadiServer, AkonadiClient, CyrusServer). This allows
us to move all fixtures into conftest.py.

Implementation of the AkonadiClient has been added, it's just a wrapper
around the 'akonadiclient' CLI tool. The tool has gained support for JSON
output for certain commands for this use-case. The format of the output
is modelled in src/akonadi/model using Pydantic.